Made setVariables.cpp
- Made includes/setVariables.cpp
- Added include statement to:
analysis/advanced-analyses/mH-scaling/scaling-code/sum_Nparticles.C
analysis/advanced-analyses/mH-scaling/scaling-code/scale_All.cpp
analysis/basic-analysis/plotting/plot_Stacks.C
analysis/basic-analysis/sum-particles/sum_Nparticles.C
