+++ analysis/advanced-analyses/mH-scaling/scaling-code/scale_All.cpp
@@ -1,46 +1,30 @@
-#include "../../includes/scaleHist.cpp"
+#include "../../../includes/scaleHist.cpp"
+#include "../../../includes/setVariables.cpp"
 
 using namespace std;
 
-vector<string> set_variables(string particle, vector<string>& input_list){
-   //Function to set the variable list based on what particle you are given
-   std::vector<std::string> v_list = input_list;
-   if (particle == "missingET_NonInt" || particle == "missingET"){
-      //For MET, no variables (i.e. the list only has one element - an empty string)
-      v_list.assign(v_list.begin(), v_list.begin()+1);
-      }
-   else if (particle == "u" || particle == "e"){
-      //For electrons and muons, do not use mass
-      v_list.assign(v_list.begin()+1, v_list.end()-1);
-   }
-   else {
-      // For all other particles
-      v_list.assign(v_list.begin()+1, v_list.end());
-   }
-
-   return v_list;
-}
-
 void scale_All()
 {
-    /*
-    Function to take files with variables for each particle type and scale these histograms according to a varying normalization factor
-    Initialize file paths, lists of particles, variables, masses, etc. to sum over
-    Initialize input and output file
-    Loop over mH masses
-        Use set_variables() function to define what variables are relevant to that particle
-        Loop over different particles
-            Initialize a sum histogram
-            Loop over different variables
-                Initialize histogram name (depends on variable, particle, number of that particle)
-                Apply scaleHist.cpp function
-    Close files
-    */
+   /*
+   Function to take files with variables for each particle type and scale these histograms according to a varying normalization factor
 
-   //Initialize some variables
+   Initialize file paths, lists of particles, variables, masses, etc. to sum over
+   Initialize input and output file
+   Loop over mH masses
+      Use setVariables() to define what variables are relevant to that particle
+      Loop over different particles
+         Initialize a sum histogram
+         Loop over different variables
+               Initialize histogram name (depends on variable, particle, number of that particle)
+               Apply scaleHist.cpp function (saves the scaled hist to the output file)
+   Close files
+   */
+
+   //Set paths to input and output files
    string f_path = "/Users/matt/Documents/work/0-CERN-UJ-HEP/particle/hZdZd/Zd-code/git_Zd/"; //Working dir
    string inFile_S = f_path + "tests_2024-07-03/summed-Particles_all-mH.root"; //In kinematics file
-   string outFile_S = f_path + "tests_2024-07-03/scaled-Particles_all-mH.root"; //Out scaled kinematics file
+   string outFile_S = f_path + "tests_2024-08-21/scaled-Particles_all-mH.root"; //Out scaled kinematics file
+   //Initialize some variables
    string particles[] = {"H", "S", "Zd", "e", "u", "missingET", "missingET_NonInt"}; //Different kinds of particles included in the input file
    float norm_factors[11] = {0.068729566, 0.221057757, 0.237594439, 0.169728543, 0.109833669, 0.069514033, 0.04661499, 0.030499267, 0.021510573, 0.014793756, 0.010123407};
    std::vector<std::string> variables = {"", "phi_", "eta_", "pT_", "m"};
