+++ analysis/advanced-analyses/mH-scaling/scaling-code/sum_Nparticles.C
@@ -1,46 +1,47 @@
-#include "../../includes/sumHists.cpp"
+#include "../../../includes/sumHists.cpp"
+#include "../../../includes/setVariables.cpp"
 
 using namespace std;
 
-vector<string> set_variables(string particle, vector<string>& input_list){
-   //Function to set the variable list based on what particle you are given
-   std::vector<std::string> v_list = input_list;
-   if (particle == "missingET_NonInt" || particle == "missingET"){
-      //For MET, no variables (i.e. the list only has one element - an empty string)
-      v_list.assign(v_list.begin(), v_list.begin()+1);
-      }
-   else if (particle == "u" || particle == "e"){
-      //For electrons and muons, do not use mass
-      v_list.assign(v_list.begin()+1, v_list.end()-1);
-   }
-   else {
-      // For all other particles
-      v_list.assign(v_list.begin()+1, v_list.end());
-   }
+// vector<string> set_variables(string particle, vector<string>& input_list){
+//    //Function to set the variable list based on what particle you are given
+//    std::vector<std::string> v_list = input_list;
+//    if (particle == "missingET_NonInt" || particle == "missingET"){
+//       //For MET, no variables (i.e. the list only has one element - an empty string)
+//       v_list.assign(v_list.begin(), v_list.begin()+1);
+//       }
+//    else if (particle == "u" || particle == "e"){
+//       //For electrons and muons, do not use mass
+//       v_list.assign(v_list.begin()+1, v_list.end()-1);
+//    }
+//    else {
+//       // For all other particles
+//       v_list.assign(v_list.begin()+1, v_list.end());
+//    }
 
-   return v_list;
-}
+//    return v_list;
+// }
 
 void sum_Nparticles()
 {
-    /*
-    Function to take analysis output files and sum the base variables for each type of particle
-    Initialize file paths, lists of particles, variables, masses, etc. to sum over
-    Initialize output file
-    Loop over different files
-        Initialize file string and TFile
-        Loop over different particles (MET is defined as its own particle)
-            Use set_variables() function to define what variables are relevant to that particle
-            Loop over different variables
-                Initialize a sum histogram
-                Loop over the number of that kind of particle
-                    Initialize histogram name (depends on variable, particle, number of that particle)
-                    Copy histogram from file
-                    Add() copy to sum histogram
-                Save sum histogram to output file
-        Close file
+   /*
+   Function to take analysis output files and sum the base variables for each type of particle
+   Initialize file paths, lists of particles, variables, masses, etc. to sum over
+   Initialize output file
+   Loop over different files
+      Initialize file string and TFile
+      Loop over different particles (MET is defined as its own particle)
+         Use setVariables() to define what variables are relevant to that particle
+         Loop over different variables
+               Initialize a sum histogram
+               Loop over the number of that kind of particle
+                  Initialize histogram name (depends on variable, particle, number of that particle)
+                  Copy histogram from file
+                  Add() copy to sum histogram
+               Save sum histogram to output file
+      Close file
    Close output file
-    */
+   */
 
    //Initialize some variables 
    string mH_masses[11] = {"300", "350", "400", "450", "500", "550", "600", "650", "700", "750", "800"}; //Different H particle masses
@@ -64,7 +65,7 @@ void sum_Nparticles()
       for (int p = 0; p < sizeof(particles)/sizeof(particles[0]); p++){
          cout << "~~~~~~~~ Getting particle: " << particles[p] << " ~~~~~~~~" << endl;
          //Set variables based on particle
-         std::vector<std::string> relevant_variables = set_variables(particles[p], variables);
+         std::vector<std::string> relevant_variables = setVariables(particles[p], variables);
 
          //Loop over variables
          for (int v = 0; v < relevant_variables.size(); v++){
