+++ analysis/basic-analysis/sum-particles/.vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "files.associations": {
+        "sum_Nparticles.C": "cpp"
+    }
+}
